Samauma @ Feature - GetTreeById & UpdateTree functions

+++ Samauma/UseCases/TreesUseCases/CreateTree/CreateTreeUseCaseInput.cs
@@ -1,10 +1,10 @@
 ﻿namespace Samauma.UseCases.CreateTree
 {
-    public class CreateTreeUseCaseInput 
+    public class CreateTreeUseCaseInput
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public int Value { get; set; }
+        public double Value { get; set; }
         public string Biome { get; set; } 
     }
 }
